# Improve performance and reduce allocations further for Plan.